+++ Assets/PlayerInputActions.inputactions
@@ -256,20 +256,64 @@
             ],
             "bindings": [
                 {
-                    "name": "",
-                    "id": "3ae1ffdb-dff4-47ee-bc48-87099f063519",
-                    "path": "",
+                    "name": "2D Vector",
+                    "id": "6d6f30de-4a3d-4e87-99d9-34d364872a7c",
+                    "path": "2DVector",
+                    "interactions": "",
+                    "processors": "",
+                    "groups": "",
+                    "action": "Move",
+                    "isComposite": true,
+                    "isPartOfComposite": false
+                },
+                {
+                    "name": "up",
+                    "id": "08ee3da1-7f1e-405a-bee7-089c207f1f2e",
+                    "path": "<Keyboard>/w",
                     "interactions": "",
                     "processors": "",
                     "groups": "",
                     "action": "Move",
                     "isComposite": false,
-                    "isPartOfComposite": false
+                    "isPartOfComposite": true
+                },
+                {
+                    "name": "down",
+                    "id": "fd4a6b82-9162-4b01-ae59-a427aab26186",
+                    "path": "<Keyboard>/s",
+                    "interactions": "",
+                    "processors": "",
+                    "groups": "",
+                    "action": "Move",
+                    "isComposite": false,
+                    "isPartOfComposite": true
+                },
+                {
+                    "name": "left",
+                    "id": "d29c6b22-0b6b-4659-8b8c-5018094e340b",
+                    "path": "<Keyboard>/a",
+                    "interactions": "",
+                    "processors": "",
+                    "groups": "",
+                    "action": "Move",
+                    "isComposite": false,
+                    "isPartOfComposite": true
+                },
+                {
+                    "name": "right",
+                    "id": "67474072-0764-48ad-bc91-42b4226cab76",
+                    "path": "<Keyboard>/d",
+                    "interactions": "",
+                    "processors": "",
+                    "groups": "",
+                    "action": "Move",
+                    "isComposite": false,
+                    "isPartOfComposite": true
                 },
                 {
                     "name": "",
                     "id": "65adb8de-b4a8-4477-98dc-d2739543ac40",
-                    "path": "",
+                    "path": "<Keyboard>/enter",
                     "interactions": "",
                     "processors": "",
                     "groups": "",
@@ -280,7 +324,7 @@
                 {
                     "name": "",
                     "id": "477a9e27-291f-4237-bbb7-57e6976db796",
-                    "path": "",
+                    "path": "<Keyboard>/escape",
                     "interactions": "",
                     "processors": "",
                     "groups": "",
