+++ Assets/Script/UI/Button/UIInput.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 1c4be6a3e3b3404db281cc12b3d5fd67
+guid: 97b5f7a3562d4c8fb11f7b37e0f4b532
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
